+++ Changelog.md
@@ -6,7 +6,7 @@
 
 ### Fixed
 
--   Programming Modem Firmware to thingy:91 no longer shows invalid warning when
+-   Programming Modem Firmware to Thingy:91 no longer shows invalid warning when
     modem firmware is the expected one
 -   Show the device family in the Memory layout title for MCU Boot devices